[P2P-104] refactor: LikeSerivce로 인터페이스 수정
