+++ src/main/java/com/prgrms/p2p/domain/like/service/LikeService.java
@@ -1,6 +1,6 @@
 package com.prgrms.p2p.domain.like.service;
 
-public interface LikePolicy {
+public interface LikeService {
 
   void toggle(Long userId, Long targetId);
 }
